requirements.txt
# ClarityHire — First-Time Job Seeker Assistant
# Dependencies

# Core LLM & Agent
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
langgraph>=0.1.0

# Vector DB & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# UI
streamlit>=1.30.0

# Evaluation
ragas>=0.1.0
datasets>=2.14.0

# Utilities
python-dotenv>=1.0.0
